made new module for basis-function-based methods, changed imports, reran jupyter notebooks
