+++ docs/source/basis_fit.rst
@@ -1,0 +1,5 @@
+basis_fit
+=========
+
+.. automodule:: pynumdiff.basis_fit
+    :members:
