+++ pynumdiff/__init__.py
@@ -7,4 +7,5 @@ from .smooth_finite_difference import meandiff, mediandiff, gaussiandiff, friedr
 from .polynomial_fit import splinediff, polydiff, savgoldiff
 from .total_variation_regularization import tvrdiff, velocity, acceleration, jerk, iterative_velocity, smooth_acceleration, jerk_sliding
 from .kalman_smooth import kalman_filter, rts_smooth, rtsdiff, constant_velocity, constant_acceleration, constant_jerk
-from .linear_model import spectraldiff, lineardiff, rbfdiff
+from .basis_fit import spectraldiff, rbfdiff
+from .linear_model import lineardiff
