+++ pynumdiff/basis_fit/__init__.py
@@ -1,0 +1,5 @@
+"""Methods based on fitting basis functions to data
+"""
+from ._basis_fit import spectraldiff, rbfdiff
+
+__all__ = ['spectraldiff', 'rbfdiff'] # So automodule from the .rst finds them
